+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Editing:
     * Action to paste as a composition
     * Action to import animations
+* UI:
+    * Right clicking on the layer view to get a context menu no longer changes the selection
+    * Object context menus now have actions to toggle visibility and lock
 
 ## 0.3.2